Add files via upload
This is working project version with options to register and to update reservation forms. MVC, DI, Composer, Interface and Exceptions used. Also with functioning validation only on initial entry of phone number (only checks the count of digits) and visit time (if not chosen and sent as empty). Still no validation of $_POST inputs.